fix: 퀴즈 페이지 상단 nav 제거
